Move NonlinearMaximizer classes into utils
