+++ src/utils/NonlinearMaximizer.cpp
